TEAM_CV/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pillow>=10.0
numpy>=1.24
opencv-python>=4.9
pydantic>=2.0
python-multipart>=0.0.9
ultralytics>=8.2.0